# packages/create-h4ckath0n/templates/fullstack/api/openapi.json
{
  "components": {
    "schemas": {
      "DeviceBindingResponse": {
        "properties": {
          "device_id": {
            "description": "Device ID that starts with the d prefix, empty when no device key is bound.",
            "title": "Device Id",
            "type": "string"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional display name for the user.",
            "title": "Display Name"
          },
          "role": {
            "description": "Server-side role for the user.",
            "title": "Role",
            "type": "string"
          },
          "user_id": {
            "description": "User ID that starts with the u prefix.",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "device_id",
          "role"
        ],
        "title": "DeviceBindingResponse",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "Standard error envelope for auth routes.",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              }
            ],
            "description": "Error detail message or structured error payload.",
            "title": "Detail"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "device_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional label for the device.",
            "title": "Device Label"
          },
          "device_public_key_jwk": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional device public key in JWK format to bind a device identity.",
            "title": "Device Public Key Jwk"
          },
          "email": {
            "description": "Account email for password-based login.",
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "description": "Plaintext password to verify.",
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "MessageResponse": {
        "properties": {
          "message": {
            "description": "Human-readable response message.",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "MessageResponse",
        "type": "object"
      },
...
      "PasswordResetConfirmSchema": {
        "properties": {
          "device_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional label for the device.",
            "title": "Device Label"
          },
          "device_public_key_jwk": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional device public key in JWK format to bind a device identity.",
            "title": "Device Public Key Jwk"
          },
          "new_password": {
            "description": "New password to set for the account.",
            "title": "New Password",
            "type": "string"
          },
          "token": {
            "description": "Password reset token issued by the server.",
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token",
          "new_password"
        ],
        "title": "PasswordResetConfirmSchema",
        "type": "object"
      },
      "PasswordResetRequestSchema": {
        "properties": {
          "email": {
            "description": "Account email to send a reset token.",
            "format": "email",
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "PasswordResetRequestSchema",
        "type": "object"
      },
      "RegisterRequest": {
        "properties": {
          "device_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional label for the device.",
            "title": "Device Label"
          },
          "device_public_key_jwk": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional device public key in JWK format to bind a device identity.",
            "title": "Device Public Key Jwk"
          },
          "display_name": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional display name for the user.",
            "title": "Display Name"
          },
          "email": {
            "description": "Account email for password-based signup.",
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "description": "Plaintext password, hashed server-side.",
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "RegisterRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "DeviceJWT": {
        "description": "Device-signed ES256 JWT minted by the client. The JWT header must include kid set to the device id, and the aud claim must be h4ckath0n:http.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Hackathon app powered by h4ckath0n",
    "title": "h4ckath0n",
    "version": "0.1.9"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "description": "Verify email and password, then bind an optional device key.",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceBindingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid email or password."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login with password",
        "tags": [
          "password-auth",
          "password-auth"
        ]
      }
    },
...
    "/auth/password-reset/confirm": {
      "post": {
        "description": "Confirm a password reset token, set a new password, and bind an optional device key.",
        "operationId": "password_reset_confirm_auth_password_reset_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetConfirmSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceBindingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid or expired reset token."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm password reset",
        "tags": [
          "password-auth",
          "password-auth"
        ]
      }
    },
    "/auth/password-reset/request": {
      "post": {
        "description": "Request a password reset token for the account. Returns the same message even when the email is unknown.",
        "operationId": "password_reset_request_auth_password_reset_request_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetRequestSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Request a password reset",
        "tags": [
          "password-auth",
          "password-auth"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "description": "Create a new account using email and password, then bind an optional device key.",
        "operationId": "register_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceBindingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request or registration error."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Email already registered."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register with password",
        "tags": [
          "password-auth",
          "password-auth"
        ]
      }
    },
...
  }
}